Revert "class diagram update"
This reverts commit 64c30f608898ec197ff20091feec80966d26a58c.

+++ C#/SEP3/FileData/DAOs/OrderFileDao.cs
@@ -15,7 +15,7 @@ public class OrderFileDao : IOrderDao
 
     public Task<Order> CreateAsync(Order order)
     {
-        long orderId = 1;
+        int orderId = 1;
         if (context.Orders.Any())
         {
             orderId = context.Orders.Max(c => c.Id);
